SW_2_0-28 Validar Eventos en los Diferentes Filtros, Inputs y Selects del Formulario para Garantizar el Correcto Funcionamiento

+++ lib/src/modules/package/factura/data/datasorces/datatables/documentos_table_data_builder.dart
@@ -111,8 +111,8 @@ class DocumentosTableDataBuilder {
   static List<PlutoRow> buildDataRows(List<Documento> documentos, BuildContext context) {
     final List<PlutoRow> dataRows = [];
     documentos.asMap().forEach((index, remesa) {
-      int totalAdiciones = remesa.adiciones.fold(0, (total, adicion) => total + adicion.valor);
-      int totalDescuentos = remesa.descuentos.fold(0, (total, descuento) => total + descuento.valor);
+      double totalAdiciones = remesa.adiciones.fold(0, (total, adicion) => total + adicion.valor);
+      double totalDescuentos = remesa.descuentos.fold(0, (total, descuento) => total + descuento.valor);
 
       Map<String, String> infoRemesa = {
         'remesa': "${remesa.impreso} (${remesa.remesa})",
