+++ lib/src/modules/package/factura/domain/entities/descuento.dart
@@ -1,7 +1,7 @@
 class Descuento {
   final int codigo;
   final String tipo;
-  final int valor;
+  final double valor;
   Descuento({
     required this.codigo,
     required this.tipo,
